Move get_mount_paths from core.util to core.path
Add tests too.
