+++ pyanaconda/core/util.py
@@ -661,14 +661,6 @@ def detect_unsupported_hardware():
     return warnings
 
 
-def get_mount_paths(devnode):
-    '''given a device node, return a list of all active mountpoints.'''
-    devno = os.stat(devnode).st_rdev
-    majmin = "%d:%d" % (os.major(devno), os.minor(devno))
-    mountinfo = (line.split() for line in open("/proc/self/mountinfo"))
-    return [info[4] for info in mountinfo if info[2] == majmin]
-
-
 def xprogressive_delay():
     """ A delay generator, the delay starts short and gets longer
         as the internal counter increases.
